añadido req de especificar branch a testear en bamboo_launch

+++ lib/tasks/test.rake
@@ -32,10 +32,11 @@ end
   desc "Lanza las tasks necesarias para ejecutar todos los tests en bamboo"
   task :bamboo_launch do
     raise "NO" unless `hostname`.strip == 'balrog'
+    raise "branch not specified!" unless ENV['BAMBOO_BRANCH']
     `rm -r #{RAILS_ROOT}/coverage/*` if File.exists?("#{RAILS_ROOT}/coverage")
     `rm -r #{RAILS_ROOT}/public/storage/*` if File.exists?("#{RAILS_ROOT}/public/storage")
-    `git branch --track staging origin/staging`
-    `git checkout staging`
+    `git branch --track #{ENV['BAMBOO_BRANCH']} origin/#{ENV['BAMBOO_BRANCH']}`
+    `git checkout #{ENV['BAMBOO_BRANCH']}`
     `git submodule init`
     `git submodule update`
     `git pull origin`
